third_party/blink/web_tests/platform/generic/svg/animations/animate-elem-07-t-drt-expected.txt

SVG 1.1 dynamic animation tests

A copy of the corresponding W3C-SVG-1.1 test, which dumps the animation at certain times

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getTransformToElement(rootSVGElement, path).e is -15
PASS getTransformToElement(rootSVGElement, path).f is -43
FAIL getTransformToElement(rootSVGElement, path).e should be close to -38.5. Was -38.36632537841797.
PASS getTransformToElement(rootSVGElement, path).f is -30
PASS getTransformToElement(rootSVGElement, path).e is -65
PASS getTransformToElement(rootSVGElement, path).f is -33
PASS getTransformToElement(rootSVGElement, path).e is -65
PASS getTransformToElement(rootSVGElement, path).f is -33
PASS successfullyParsed is true

TEST COMPLETE
